requirements.txt
SpeechRecognition==3.10.0
python-dotenv==1.0.0
pydub==0.25.1
sounddevice==0.4.6
ollama==0.1.6
PyAudio==0.2.14 